Rename Source services module into LMS services module
