+++ src/langium-model-server/langium-model-server-module.ts
@@ -5,8 +5,8 @@ import { LmsRenameProvider } from './lsp/rename-provider'
 import type { SemanticIdentity } from './semantic/identity'
 import type { IdentityIndex } from './semantic/identity-index'
 import type { LangiumModelServerDefaultServices, LangiumModelServerDefaultSharedServices, LangiumModelServerServices } from './services'
-import { DefaultLangiumSourceModelServer } from './source/source-model-server'
-import { DefaultSourceModelSubscriptions } from './source/source-model-subscriptions'
+import { DefaultLangiumSourceModelServer } from './lms/source-model-server'
+import { DefaultSourceModelSubscriptions } from './lms/source-model-subscriptions'
 import type { LmsDocument } from './workspace/documents'
 import { DefaultLmsDocumentBuilder } from './workspace/lms-document-builder'
 
@@ -21,7 +21,7 @@ Module<LangiumModelServerServices<SM, II, D>, LangiumModelServerDefaultServices>
         workspace: {
             LmsDocumentBuilder: (services) => new DefaultLmsDocumentBuilder(services),
         },
-        source: {
+        lms: {
             LangiumSourceModelServer: (services) => new DefaultLangiumSourceModelServer(services),
             SourceModelSubscriptions: () => new DefaultSourceModelSubscriptions()
         }
