+++ src/langium-model-server/lms/model/action/highlight.ts
