+++ src/langium-model-server/lms/model/action/rename.ts
