+++ src/langium-model-server/lms/model/highlight-response.ts
