+++ src/language-server/task-list/lms/task-list-source-update-combiner.ts
@@ -1,8 +1,8 @@
 import { stream } from 'langium'
 import type * as id from '../../../langium-model-server/semantic/identity'
-import { ArrayUpdate, Update } from '../../../langium-model-server/source/model'
-import type { ElementAttributes, ElementUpdate } from '../../../langium-model-server/source/model'
-import type { SourceUpdateCombiner } from '../../../langium-model-server/source/source-update-combiner'
+import { ArrayUpdate, Update } from '../../../langium-model-server/lms/model'
+import type { ElementAttributes, ElementUpdate } from '../../../langium-model-server/lms/model'
+import type { SourceUpdateCombiner } from '../../../langium-model-server/lms/source-update-combiner'
 import type { Model, Task, Transition } from './model'
 
 export class TaskListSourceUpdateCombiner implements SourceUpdateCombiner<Model> {
