+++ src/language-server/task-list/semantic/task-list-identity-manager.ts
@@ -1,6 +1,6 @@
 import type { ModelExposedIdentityIndex } from '../../../langium-model-server/semantic/identity-index'
 import { AbstractIdentityManager } from '../../../langium-model-server/semantic/identity-manager'
-import type * as source from '../source/model'
+import type * as source from '../lms/model'
 import type { TaskListDocument } from '../workspace/documents'
 import { Model } from './task-list-identity'
 import { TaskListIdentityIndex } from './task-list-identity-index'
