+++ src/language-server/task-list/semantic/task-list-identity-storage.ts
@@ -3,7 +3,7 @@ import { AbstractIdentityStorage } from '../../../langium-model-server/semantic/
 import type { TaskListDocument } from '../workspace/documents'
 import { Model } from './task-list-identity'
 import type { TaskListIdentityIndex } from './task-list-identity-index'
-import type * as source from '../source/model'
+import type * as source from '../lms/model'
 
 export class TaskListIdentityStorage extends AbstractIdentityStorage<source.Model, TaskListIdentityIndex, TaskListDocument> implements IdentityStorage {
