+++ src/language-server/task-list/task-list-module.ts
@@ -6,10 +6,10 @@ import { TaskListIdentityManager } from './semantic/task-list-identity-manager'
 import { TaskListIdentityReconciler } from './semantic/task-list-identity-reconciler'
 import { TaskListIdentityStorage } from './semantic/task-list-identity-storage'
 import { TaskListSemanticDomain } from './semantic/task-list-semantic-domain'
-import type * as source from './source/model'
-import { TaskListSourceModelService } from './source/task-list-source-model-service'
-import { TaskListSourceUpdateCombiner } from './source/task-list-source-update-combiner'
-import { TaskListSourceUpdateManager } from './source/task-list-source-update-manager'
+import type * as source from './lms/model'
+import { TaskListSourceModelService } from './lms/task-list-source-model-service'
+import { TaskListSourceUpdateCombiner } from './lms/task-list-source-update-combiner'
+import { TaskListSourceUpdateManager } from './lms/task-list-source-update-manager'
 import type { TaskListDocument } from './workspace/documents'
 import { isTaskListDocument } from './workspace/documents'
 
@@ -24,7 +24,7 @@ export type TaskListAddedServices = {
         // Redefining the type of IdentityManager to be used in TaskListIdentityReconciler
         IdentityManager: TaskListIdentityManager
     },
-    source: {
+    lms: {
         // Redefining the type
         SourceUpdateManager: TaskListSourceUpdateManager
     }
@@ -54,7 +54,7 @@ export const TaskListModule: Module<TaskListServices, PartialLangiumModelServerS
         IdentityReconciler: (services) => new TaskListIdentityReconciler(services),
         SemanticDomainFactory: () => TaskListSemanticDomain.create,
     },
-    source: {
+    lms: {
         SourceModelService: (services) => new TaskListSourceModelService(services),
         SourceUpdateManager: () => new TaskListSourceUpdateManager(),
         SourceUpdateCombiner: () => new TaskListSourceUpdateCombiner(),
